requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Web Scraping & HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration & Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Excel Export
xlsxwriter==3.1.9

